website/requirements.txt
asgiref==3.5.2
boto3==1.24.39
botocore==1.27.39
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
cryptography==37.0.4
defusedxml==0.7.1
distlib==0.3.5
Django==3.2.14
django-admin-honeypot==1.1.0
django-allauth==0.51.0
django-cities==0.6
django-ckeditor==6.3.2
django-cors-headers==3.13.0
django-debug-toolbar==3.5.0
django-js-asset==2.0.0
django-model-utils==4.2.0
django-paypal==2.0
django-storages==1.12.3
Django-Verify-Email==2.0.3
idna==3.3
jmespath==1.0.1
oauthlib==3.2.0
Pillow==9.2.0
pycparser==2.21
PyJWT==2.4.0
python-dateutil==2.8.2
python-dotenv==0.20.0
python3-openid==3.2.0
pytz==2022.1
requests==2.28.1
requests-oauthlib==1.3.1
s3transfer==0.6.0
six==1.16.0
sqlparse==0.4.2
stripe==3.5.0
swapper==1.3.0
tqdm==4.64.0
typing_extensions==4.3.0
urllib3==1.26.9
uuid==1.30
pymysql ==1.0.2
django_advanced_password_validation == 1.1.1